move dataflow-graph pieces to core
