+++ javascript/core/src/graph/__tests__/NodeImpl.spec.ts
