+++ javascript/core/src/graph/__tests__/VariadicNodeImpl.spec.ts
