--- javascript/dataflow-graph/docs/dataflow-graph.api.md
@@ -1,77 +1,0 @@
-## API Report File for "@data-wrangling-components/dataflow-graph"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import type { Observable } from 'rxjs';
-
-// @public
-export type Maybe<T> = T | undefined;
-
-// @public
-interface Node_2<T, Config = unknown> {
-    config: Maybe<Config>;
-    readonly inputs: string[];
-    install(name: string, socket: Observable<Maybe<T>>): void;
-    output(name?: string): Observable<Maybe<T>>;
-    readonly outputs: string[];
-    outputValue(name?: string): Maybe<T>;
-    uninstall(name: string): void;
-}
-export { Node_2 as Node }
-
-// Warning: (ae-missing-release-tag) "NodeImpl" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
-//
-// @public (undocumented)
-export abstract class NodeImpl<T, Config> implements Node_2<T, Config> {
-    constructor(inputs?: string[], outputs?: string[]);
-    // (undocumented)
-    get config(): Maybe<Config>;
-    set config(value: Maybe<Config>);
-    // (undocumented)
-    protected abstract doRecalculate(): Promise<void> | void;
-    protected emit(value: Maybe<T>, output?: string): void;
-    // (undocumented)
-    protected emitError(error: unknown): void;
-    // (undocumented)
-    readonly inputs: string[];
-    // (undocumented)
-    protected inputValue(name: string): Maybe<T>;
-    // (undocumented)
-    protected get inputValues(): Map<string, Maybe<T>>;
-    // (undocumented)
-    install(name: string, socket: Observable<Maybe<T>>): void;
-    // (undocumented)
-    output(name?: string): Observable<Maybe<T>>;
-    // (undocumented)
-    readonly outputs: string[];
-    // (undocumented)
-    outputValue(name?: string): Maybe<T>;
-    protected recalculate(): Promise<void>;
-    // (undocumented)
-    uninstall(name: string): void;
-    // (undocumented)
-    protected verifyInputSocketName(name: string): void;
-    // (undocumented)
-    protected verifyOutputSocketName(name: string): void;
-}
-
-// Warning: (ae-missing-release-tag) "VariadicNodeImpl" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
-//
-// @public (undocumented)
-export abstract class VariadicNodeImpl<T, Config> extends NodeImpl<T, Config> {
-    constructor(inputs?: string[], outputs?: string[]);
-    // (undocumented)
-    protected getVariadicInputValues(): Maybe<T>[];
-    // (undocumented)
-    installNext(socket: Observable<Maybe<T>>): string;
-    // (undocumented)
-    nextInput(): string;
-    // (undocumented)
-    protected verifyInputSocketName(name: string): void;
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
