--- javascript/dataflow-graph/package.json
@@ -1,23 +1,0 @@
-{
-	"name": "@data-wrangling-components/dataflow-graph",
-	"version": "0.0.1",
-	"type": "module",
-	"main": "src/index.ts",
-	"publishConfig": {
-		"main": "dist/lib/index.js",
-		"types": "dist/lib/index.d.ts"
-	},
-	"sideEffects": false,
-	"scripts": {
-		"clean": "essex clean dist lib",
-		"build": "essex build --docs"
-	},
-	"devDependencies": {
-		"@essex/scripts": "^21.0.1",
-		"@essex/tsconfig-base": "^1.0.2",
-		"@types/jest": "^27.4.1"
-	},
-	"dependencies": {
-		"rxjs": "^7.5.5"
-	}
-}
