--- javascript/dataflow-graph/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"extends": "@essex/tsconfig-base",
-	"compilerOptions": {
-		"target": "ES2020",
-		"module": "ES2020",
-		"moduleResolution": "Node",
-		"lib": ["ESNext", "DOM", "DOM.Iterable", "ES2020.BigInt"],
-		"skipLibCheck": true,
-		"skipDefaultLibCheck": true
-	},
-	"include": ["src"]
-}
